Fix MSVC not compiling

+++ src/socket.cpp
@@ -312,7 +312,7 @@ bool UDPSocket::WaitData(int timeout_ms)
 
 	// Initialize time out struct
 	struct timeval tv;
-	tv.tv_sec = floor(timeout_ms / 1000) ;
+	tv.tv_sec = floor((float)timeout_ms / 1000) ;
 	tv.tv_usec = 1000 * (timeout_ms % 1000);
 	// select()
 	result = select(m_handle+1, &readset, NULL, NULL, &tv);
@@ -574,7 +574,7 @@ bool TCPSocket::WaitData(int timeout_ms)
 
 	// Initialize time out struct
 	struct timeval tv;
-	tv.tv_sec = floor(timeout_ms / 1000);
+	tv.tv_sec = floor((float)timeout_ms / 1000);
 	tv.tv_usec = 1000 * (timeout_ms % 1000);
 
 	if (m_bstart != m_bend)
